+++ BTS.Web/Models/ProfileViewModel.cs
@@ -53,11 +53,11 @@ namespace BTS.Web.Models
         [Display(Name = "Phí Kiểm định")]
         [DisplayFormat(DataFormatString = "{0:N0}", ApplyFormatInEditMode = true)]
         [Required(AllowEmptyStrings = false, ErrorMessage = "Yêu cầu nhập Số Phí Kiểm định")]
-        [Range(1, int.MaxValue, ErrorMessage = "Yêu cầu nhập Phí Kiểm định là số nguyên trong phạm vi [1->2147483647]")]
+        [Range(1, long.MaxValue, ErrorMessage = "Yêu cầu nhập Phí Kiểm định là số nguyên trong phạm vi [1->2147483647]")]
         [RegularExpression(@"[1-9][0-9]*$", ErrorMessage = "Yêu cầu nhập Số Trạm BTS tiếp nhận là số nguyên")]
         //Ap dung cho so Double
         //[RegularExpression(@"^\d+.\d{0,}$", ErrorMessage = "Yêu cầu nhập Số Phí Kiểm định là số nguyên")]
-        public int Fee { get; set; }
+        public long Fee { get; set; }
 
         [Display(Name = "Số CV Báo phí")]
         [StringLength(30, ErrorMessage = "Số CV Báo phí không quá 30 ký tự")]
